requirements.txt
dash==2.16.1
numpy==1.26.4
opencv_python==4.9.0.80
Pillow==10.2.0
gunicorn
pandas
dash-tools
